separate logic out of componants to containers
