+++ app/src/components/map/index.js
@@ -1,19 +1,10 @@
-import React, { Component } from 'react'
+import React from 'react'
 import ReactMapboxGl, { Layer, Feature } from 'react-mapbox-gl'
-import config from '../../config/base'
 
-const mapboxConfig = config.mapbox
-const accessToken = mapboxConfig.accessToken
-const Map = ReactMapboxGl({ accessToken: accessToken })
-
-const MapComponent = ({ lat, lon }) => {
+const MapComponent = ({ lat, lon, accessToken, containerStyle, style, zoom }) => {
+  const Map = ReactMapboxGl({ accessToken: accessToken })
   return (
-    <Map
-      containerStyle={mapboxConfig.containerStyle}
-      style={mapboxConfig.style}
-      center={[lat, lon]}
-      zoom={[mapboxConfig.zoom]} className="mx-2"
-    >
+    <Map containerStyle={containerStyle} style={style} center={[lat, lon]} zoom={[zoom]} className="mx-2">
       <Layer type="symbol" id="marker" layout={{ 'icon-image': 'marker-15' }}>
         <Feature coordinates={[lat, lon]} />
       </Layer>
@@ -21,32 +12,4 @@ const MapComponent = ({ lat, lon }) => {
   )
 }
 
-class MapHandler extends Component {
-  constructor(props) {
-    super(props)
-    this.state = {}
-  }
-
-  addressToGeoCode = address => {
-    const url = `https://api.mapbox.com/geocoding/v5/mapbox.places/${address}.json?access_token=${accessToken}`
-    return fetch(url).then(response => response.json())
-  }
-
-  componentDidMount() {
-    this.addressToGeoCode(this.props.address).then(response => {
-      this.setState({
-        lat: response.features[0].center[0],
-        lon: response.features[0].center[1]
-      })
-    })
-  }
-
-  render() {
-    if (!this.state.lat) {
-      return null
-    }
-    return <MapComponent lat={this.state.lat} lon={this.state.lon} />
-  }
-}
-
-export default MapHandler
+export default MapComponent
